feat: :sparkles: Move controls into a separate component
